since LocoNetBus has nothing to do with Loconet now, it's renamed
